Factored out exhibit fetch
