+++ src/actions/action-types.js
@@ -2,6 +2,7 @@
 // if this gets unweildy, it can be split into multiple files
 
 // Map preview
+export const EXHIBIT_CACHE_SAVE = 'EXHIBIT_CACHE_SAVE';
 export const RECORD_CACHE_UPDATE = 'RECORD_CACHE_UPDATE';
 export const PREVIEW_FILLCOLOR = 'PREVIEW_FILLCOLOR';
 export const PREVIEW_BASELAYER = 'PREVIEW_BASELAYER';
@@ -16,6 +17,8 @@ export const HAS_UNSAVED_CHANGES = 'HAS_UNSAVED_CHANGES';
 export const EXHIBIT_CACHE_UPDATE = 'EXHIBIT_CACHE_UPDATE';
 export const EXHIBIT_LOADING = 'exhibitShow/EXHIBIT_LOADING';
 export const EXHIBIT_ERRORED = 'exhibitShow/EXHIBIT_ERRORED';
+export const EXHIBIT_FETCH = 'EXHIBIT_FETCH';
+export const EXHIBIT_FETCH_RESPONSE_RECEIVED = 'EXHIBIT_FETCH_RESPONSE_RECEIVED';
 export const EXHIBIT_FETCH_SUCCESS = 'exhibitShow/EXHIBIT_FETCH_SUCCESS';
 export const EXHIBIT_RESET = 'exhibitShow/EXHIBIT_RESET';
 export const EXHIBIT_LOADED = 'exhibitShow/EXHIBIT_LOADED';
@@ -34,6 +37,7 @@ export const RECORD_REPLACED = 'exhibitShow/RECORD_REPLACED';
 export const RECORD_REMOVED = 'exhibitShow/RECORD_REMOVED';
 export const RECORD_COVERAGE_SET = 'exhibitShow/RECORD_COVERAGE_SET';
 
+
 // Records
 export const RECORD_ERROR = 'RECORD_ERROR';
 export const NEW_RECORD_RESET = 'recordCreate/NEW_RECORD_RESET';
@@ -50,3 +54,19 @@ export const UPDATE_RECORD_RESPONSE_RECEIVED = 'UPDATE_RECORD_RESPONSE_RECEIVED'
 export const EXHIBIT_UPDATED = 'exhibitUpdate/EXHIBIT_CREATED';
 export const EXHIBIT_PATCH_SUCCESS = 'exhibitUpdate/EXHIBIT_POST_SUCCESS';
 export const EXHIBIT_PATCH_ERRORED = 'exhibitUpdate/EXHIBIT_POST_ERRORED';
+
+
+
+// Exhibits
+export const EXHIBIT_CREATED = 'exhibitCreate/EXHIBIT_CREATED';
+export const EXHIBIT_POST_SUCCESS = 'exhibitCreate/EXHIBIT_POST_SUCCESS';
+export const EXHIBIT_POST_ERRORED = 'exhibitCreate/EXHIBIT_POST_ERRORED';
+export const NEW_EXHIBIT_RESET = 'exhibitCreate/NEW_EXHIBIT_RESET';
+
+export const EXHIBIT_DELETED = 'exhibitDelete/EXHIBIT_DELETED';
+export const EXHIBIT_DELETE_SUCCESS = 'exhibitDelete/EXHIBIT_DELETE_SUCCESS';
+export const EXHIBIT_DELETE_ERRORED = 'exhibitDelete/EXHIBIT_DELETE_ERRORED';
+
+export const EXHIBITS_LOADING = 'exhibits/EXHIBITS_LOADING';
+export const EXHIBITS_ERRORED = 'exhibits/EXHIBITS_ERRORED';
+export const EXHIBITS_FETCH_SUCCESS = 'exhibits/EXHIBITS_FETCH_SUCCESS';
