+++ src/i18nLibrary.js
@@ -1,4 +1,4 @@
-import React, {Component} from 'react';
+//import React, {Component} from 'react';
 import LocalizedStrings from 'react-localization';
 
 export let strings = new LocalizedStrings({
