+++ src/reducers/exhibit-initialState.js
@@ -1,0 +1,5 @@
+export default {
+  exhibits: [],
+  loading: false,
+  errored: false
+};
