+++ src/reducers/exhibitPreview-initialState.js
@@ -1,0 +1,3 @@
+export default {
+	cache:[]
+};
