+++ src/reducers/index.js
@@ -2,24 +2,31 @@ import { combineReducers } from 'redux';
 import { routerReducer } from 'react-router-redux';
 import { reducer as formReducer } from 'redux-form';
 import user from './not_refactored/user';
-import exhibits from './not_refactored/exhibits';
+
 import exhibitShow from './not_refactored/exhibitShow';
-import exhibitCreate from './not_refactored/exhibitCreate';
 import exhibitUpdate from './not_refactored/exhibitUpdate';
+/*
+import exhibits from './not_refactored/exhibits';
+import exhibitCreate from './not_refactored/exhibitCreate';
+*/
 import exhibitDelete from './not_refactored/exhibitDelete';
+
 import recordMapLayers from './not_refactored/recordMapLayers';
 
-import recordReducer from './reducer-record.js';
-import mapPreviewReducer from './reducer-mapPreview.js';
-import exhibitPreviewReducer from './reducer-exhibitPreview.js';
+
+import exhibitReducer from './exhibit-reducer.js';
+import recordReducer from './record-reducer.js';
+import mapPreviewReducer from './mapPreview-reducer.js';
+import exhibitPreviewReducer from './exhibitPreview-reducer.js';
 
 export default combineReducers({
   routing: routerReducer,
   form: formReducer,
   user,
-  exhibits,
+  exhibits:exhibitReducer,
+  //exhibits,
   exhibitShow,
-  exhibitCreate,
+  //exhibitCreate,
   exhibitUpdate,
   exhibitDelete,
   record: recordReducer,
