--- src/reducers/initialState-exhibitPreview.js
@@ -1,3 +1,0 @@
-export default {
-	cache:[]
-};
