--- src/reducers/initialState-record.js
@@ -1,6 +1,0 @@
-export default {
-	changedRecord: null,
-	newRecord: null,
-	loading: false,
-	error: ''
-};
