+++ src/reducers/record-initialState.js
@@ -1,0 +1,6 @@
+export default {
+	changedRecord: null,
+	newRecord: null,
+	loading: false,
+	error: ''
+};
