+++ src/views/exhibits/update.js
@@ -7,9 +7,12 @@ import {strings} from '../../i18nLibrary';
 
 const ExhibitUpdate = props => (
   <div className="ps_n3_exhibitFormContainer">
-    <ExhibitForm exhibit={props.exhibit} onSubmit={props.submit} submitLabel={strings.save_exhibit} disabled={props.loading} />
+    <ExhibitForm 	exhibit={props.exhibit}
+					onSubmit={props.submit}
+					submitLabel={strings.save_exhibit}
+					disabled={props.loading} />
     {props.errored &&
-      <p>{strings.update_exhibit_error}</p>
+    	<p>{strings.update_exhibit_error}</p>
     }
   </div>
 )
